% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.r
\docType{data}
\name{fire}
\alias{fire}
\title{Fire data set.}
\format{
A data frame with 15 rows and 2 variables
\describe {
  \item{DISTANCE}{distance in miles from nearest fire station}
  \item{Damage}{damage caused by fire in thousands of dollars}
  ...
}
}
\source{
\url{https://bookshelf.vitalsource.com/#/books/9781498728881}
}
\usage{
fire
}
\description{
A data set showing the distance of fires from nearest fire station and the cost of their damage.
}
\keyword{datasets}
