% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Bootstrap Generator}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of times to resample}

\item{x}{sample that bootstrapping will be performed on. This sample is representative of the population}

\item{fun}{statistic to be computed on the samples}

\item{alpha}{for an N% confidence interval of the given statistic, alpha=1-(N/100)}

\item{cx}{character expansion value}

\item{...}{additional arguments for plot title}
}
\value{
creates a histogram showing the sampling distribution and 100(1-alpha)% confidence interval for the given statistic. Returns a list containing the sampling distribution, confidence interval, and the values passed to fun and x.
}
\description{
Title
}
\examples{
\dontrun{myboot2(x=sample)}
}
