% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{Mean Confidence Interval}
\usage{
myci(x)
}
\arguments{
\item{x}{sample}
}
\value{
95% confidence interval for mean
}
\description{
Title
}
\examples{
\dontrun{myci(rnorm(30, 10, 12))}
}
