% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltb.R
\name{mycltb}
\alias{mycltb}
\title{Binomial Sample Mean Distribution}
\usage{
mycltb(n, iter, p = 0.5, ...)
}
\arguments{
\item{n}{sample size}

\item{iter}{number of samples}

\item{p}{probability of success associated with binomial population}
}
\value{
histogram comparing the sample mean distribution with the theoretical normal distribution anticipated by the central limit theorem
}
\description{
Title
}
\examples{
\dontrun{mycltb(n=20,iter=10000,p=0.3)}
}
