% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Normal Distribution and Cumulative Probability Plot}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{mean of the normal distribution}

\item{sigma}{standard deviation of the normal distribution}

\item{a}{for a given a, this function calculates P(Y<=a) for the given normal distribution}
}
\value{
the cumulative probability P(Y<=a)
}
\description{
Title
}
\examples{
\dontrun{myncurve(mu=10,sigma=5, a=6)}
}
