% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mynrml.R
\name{myNRML}
\alias{myNRML}
\title{Newton Raphson Maximum Likelihood}
\usage{
myNRML(x0, delta = 0.001, llik, xrange, parameter = "param")
}
\arguments{
\item{x0}{initial x value}

\item{delta}{threshold that stops iteration above y-axis}

\item{llik}{log likelihood function}

\item{xrange}{x-limits for curve}

\item{parameter}{parameter of the log likelihood function}
}
\value{
plots the log likelihood function and newton-rapshon algorithm on the derivative
}
\description{
Title
}
\examples{
\dontrun{myNRML(1, llik=llogpois, xrange=c(0,40), parameter="lambda")}
}
