% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvreader.R
\name{myread}
\alias{myread}
\title{CSV reader}
\usage{
myread(dird, csv)
}
\arguments{
\item{dird}{the directory housing the CSV file (ending with "\\\")}

\item{csv}{the name of the CSV file}
}
\value{
a dataframe object containing the data in the csv file.
}
\description{
Title
}
\examples{
\dontrun{mpg.df=myread(paste0("C:\\\\", "CSVdirectory\\\\"), "Data.csv")}
}
