% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson.R
\name{poisson}
\alias{poisson}
\title{Poisson Probability Distribution}
\usage{
poisson(lambda, y)
}
\arguments{
\item{lambda}{mean number of events during given units of time, volume, or area.}

\item{y}{value of poisson random variable}
}
\value{
probability of y
}
\description{
Title
}
\examples{
\dontrun{prob_3=poisson(2,3)}
}
